Add common base class for image-based states
`AbstractImageTokenOverlay` is the base for `ImageTokenOverlay`, `CornerImageTokenOverlay`, and `FlowImageTokenOverlay`.
In places where a type check is needed to look for an image-based overlay, `AbstractImageTokenOverlay` is now checked
instead of `ImageTokenOverlay`. The new base class also provides an implementation of `#getAssetId()` and `#getAssets()`
so child classes don't have to worry about it.
